# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AVITO API",
        "description": "API endpoints for AVITO",
        "contact": {
            "email": "pashaboga4ev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:80/api",
            "description": "Base server for development"
        }
    ],
    "paths": {
        "/banner": {
            "get": {
                "tags": [
                    "Avito Banner"
                ],
                "summary": "Getting all banners with filtering by feature and/or tag",
                "operationId": "bannersList",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    },
                    {
                        "name": "feature_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Идентификатор фичи",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tag_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Идентификатор тега",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Limit",
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Offset",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BannerCollectionResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    }
                }
            },
            "post": {
                "tags": [
                    "Avito Banner"
                ],
                "summary": "Создание нового баннера",
                "operationId": "createBanner",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateBannerRequestBody"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/CreateBannerResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    }
                }
            }
        },
        "/banner/toggle-status": {
            "put": {
                "tags": [
                    "Avito Banner"
                ],
                "summary": "Toggle banner status",
                "operationId": "toggleStatus",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateStatusBannerRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SuccessResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    }
                }
            }
        },
        "/banner/{id}": {
            "delete": {
                "tags": [
                    "Avito Banner"
                ],
                "summary": "Удаление баннера по идентификатору",
                "operationId": "deleteBanner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/SuccessDeletedResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Avito Banner"
                ],
                "summary": "Обновление содержимого баннера",
                "operationId": "updateBanner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "Идентификатор баннера",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateBannerRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BannerResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    }
                }
            }
        },
        "/banner/bulk-delete": {
            "delete": {
                "tags": [
                    "Avito Banner"
                ],
                "summary": "Bulk delete banners",
                "operationId": "bulkDeleteBanners",
                "parameters": [
                    {
                        "name": "tag_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "feature_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SuccessResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateBannerRequest": {
                "required": [
                    "tag_ids",
                    "feature_id",
                    "content",
                    "is_active"
                ],
                "properties": {
                    "tag_ids": {
                        "description": "Идентификаторы тэгов",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    },
                    "feature_id": {
                        "description": "Идентификатор фичи",
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "description": "Содержимое баннера",
                        "type": "object",
                        "example": "{\"title\": \"some_title\", \"text\": \"some_text\", \"url\": \"some_url\"}",
                        "additionalProperties": true
                    },
                    "is_active": {
                        "description": "Флаг активности баннера",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "UpdateBannerRequest": {
                "required": [
                    "tag_ids",
                    "feature_id",
                    "content",
                    "is_active"
                ],
                "properties": {
                    "tag_ids": {
                        "description": "Идентификаторы тэгов",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ],
                        "nullable": true
                    },
                    "feature_id": {
                        "description": "Идентификатор фичи",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "content": {
                        "description": "Содержимое баннера",
                        "type": "object",
                        "example": "{\"title\": \"some_title\", \"text\": \"some_text\", \"url\": \"some_url\"}",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "is_active": {
                        "description": "Флаг активности баннера",
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateStatusBannerRequest": {
                "required": [
                    "banner_id",
                    "is_active"
                ],
                "properties": {
                    "banner_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "BannerCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/BannerResource"
                }
            },
            "BannerResource": {
                "required": [
                    "banner_id",
                    "tag_ids",
                    "feature_id",
                    "content",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "banner_id": {
                        "description": "Идентификатор баннера",
                        "type": "integer",
                        "example": 1
                    },
                    "tag_ids": {
                        "description": "Идентификаторы тэгов",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    },
                    "feature_id": {
                        "description": "Идентификатор фичи",
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "{\"title\": \"some_title\", \"text\": \"some_text\", \"url\": \"some_url\"}"
                    },
                    "is_active": {
                        "description": "Флаг активности баннера",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Дата создания баннера",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления баннера",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CreateBannerResource": {
                "required": [
                    "banner_id"
                ],
                "properties": {
                    "banner_id": {
                        "description": "Идентификатор созданного баннера",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Forbidden": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized"
                    }
                },
                "type": "object"
            },
            "InternalServerError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "InvalidData": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Any error message"
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
            "Success": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Success."
                    }
                },
                "type": "object"
            },
            "Unauthorized": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "BannerCollectionResponse": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BannerCollection"
                        }
                    }
                }
            },
            "BannerResponse": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BannerResource"
                        }
                    }
                }
            },
            "CreateBannerResponse": {
                "description": "Created",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateBannerResource"
                        }
                    }
                }
            },
...
            "ForbiddenResponse": {
                "description": "Пользователь не имеет доступа",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Forbidden"
                        }
                    }
                }
            },
            "InternalServerErrorResponse": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/InternalServerError"
                        }
                    }
                }
            },
            "InvalidDataResponse": {
                "description": "Некорректные данные",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/InvalidData"
                        }
                    }
                }
            },
            "NotFoundResponse": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFound"
                        }
                    }
                }
            },
            "SuccessResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Success"
                        }
                    }
                }
            },
            "SuccessDeletedResponse": {
                "description": "Success Deleted"
            },
            "UnauthorizedResponse": {
                "description": "Пользователь не авторизован",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Unauthorized"
                        }
                    }
                }
            },
...
        },
        "requestBodies": {
...
            "CreateBannerRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateBannerRequest"
                        }
                    }
                }
            },
            "UpdateBannerRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateBannerRequest"
                        }
                    }
                }
            },
            "UpdateStatusBannerRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateStatusBannerRequest"
                        }
                    }
                }
            },
...
        }
    },
...
}